SOC-9254: Block vkcon launching when card is not in boot2 running
If a PCIe transaction is issued to BAR2 and accessing
DDR memory while only UCODE is runnig, it causes some of the
PCIe switch to find links getting into error state.
These correctable errors would continuously flood the dmesgs.
Add a check and if boot2 is not running, there is no need to
initialize and poke BAR2.

Change-Id: I8e67cb1520c2616bff398e25edb7e584dd635274
Reviewed-on: http://gerrit-ccxsw.broadcom.net/118157
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Mircea Carausu <mircea.carausu@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>